backend/requirements/local.txt
-r base.txt

Werkzeug==2.0.2
ipdb==0.13.9
psycopg2==2.9.2
watchgod==0.7

# Testing
# ------------------------------------------------------------------------------
mypy==0.910
django-stubs==1.9.0
pytest==6.2.5
pytest-sugar==0.9.4
djangorestframework-stubs==1.4.0
cerberus==1.3.4
model-bakery==1.3.3

# Documentation
# ------------------------------------------------------------------------------
sphinx==4.3.1
sphinx-autobuild==2021.3.14

# Code quality
# ------------------------------------------------------------------------------
flake8==4.0.1
flake8-isort==4.1.1
coverage==6.2
black==21.11b1
pylint-django==2.4.4
pylint-celery==0.3
pre-commit==2.16.0

# Django
# ------------------------------------------------------------------------------
factory-boy==3.2.1

django-debug-toolbar==3.2.2
django-extensions==3.1.5
django-coverage-plugin==2.0.2
pytest-django==4.5.1
